change position of SportsmanExtend
